% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exgaussian2_stancode.R
\name{exgaussian2_stancode}
\alias{exgaussian2_stancode}
\title{\code{brms} \code{stanvars}  for the standard parameterization of the Ex-Gaussian distribution}
\usage{
exgaussian2_stancode(fixed_mu = FALSE, fixed_sigma = FALSE, fixed_tau = FALSE)
}
\arguments{
\item{fixed_mu}{Is \code{mu} fixed (i.e., \strong{not} modeled?)}

\item{fixed_sigma}{Is \code{sigma} fixed (i.e., \strong{not} modeled?)}

\item{fixed_tau}{Is \code{tau} fixed (i.e., \strong{not} modeled?)}
}
\value{
An object of class \code{c("stanvars")} to be used in \code{brms(stanvars=)}.
}
\description{
\code{brms} \code{stanvars}  for the standard parameterization of the Ex-Gaussian distribution
}
