% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexgaussian2.R
\name{rexgaussian2}
\alias{rexgaussian2}
\title{Generate random values from standard parameterization of the Ex-Gaussian distribution}
\usage{
rexgaussian2(n, mu, sigma, tau)
}
\arguments{
\item{n}{Number of draws to sample from the distribution.}

\item{mu}{Vector of means of the combined distribution.}

\item{sigma}{Vector of standard deviations of the gaussian component.}

\item{tau}{Vector of scales of the exponential component.}
}
\description{
Generate random values from standard parameterization of the Ex-Gaussian distribution
}
